fix: refactor to remove basedir adn rel on env var to override
