+++ cmd/alzlibtool/command/check/library.go
@@ -31,7 +31,7 @@ var libraryCmd = cobra.Command{
 		}
 		az.AddPolicyClient(cf)
 		thisRef := alzlib.NewCustomLibraryReference(args[0])
-		libs, err := alzlib.FetchAllLibrariesWithDependencies(cmd.Context(), ".alzlib", 0, thisRef, make(alzlib.LibraryReferences, 0, 5))
+		libs, err := alzlib.FetchAllLibrariesWithDependencies(cmd.Context(), 0, thisRef, make(alzlib.LibraryReferences, 0, 5))
 		if err != nil {
 			cmd.PrintErrf("%s could not fetch all libraries with dependencies: %v\n", cmd.ErrPrefix(), err)
 		}
