+++ cmd/alzlibtool/command/document/library.go
@@ -3,6 +3,7 @@ package document
 import (
 	"os"
 
+	"github.com/Azure/alzlib"
 	"github.com/Azure/alzlib/pkg/doc"
 	"github.com/spf13/cobra"
 )
@@ -13,7 +14,11 @@ var documentLibraryBaseCmd = cobra.Command{
 	Long:  `Generates documentation for the supplied library path.`,
 	Args:  cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
-		fs := os.DirFS(args[0])
-		doc.AlzlibReadmeMd(cmd.Context(), os.Stdout, fs)
+		thislib := alzlib.NewCustomLibraryReference(args[0])
+		alllibs, err := alzlib.FetchAllLibrariesWithDependencies(cmd.Context(), 0, thislib, make(alzlib.LibraryReferences, 0, 5))
+		if err != nil {
+			cmd.PrintErrf("%s could not fetch all libraries with dependencies: %v\n", cmd.ErrPrefix(), err)
+		}
+		doc.AlzlibReadmeMd(cmd.Context(), os.Stdout, alllibs.FSs()...)
 	},
 }
